Cancel coroutine in Future.cancel instead of AbstractFuture.afterDone
afterDone is still in beta. This also allowed to replace InnerFuture with SettableFuture.
Two tests are ignored because now cancellation propagates after listeners are executed. There's a TODO to fix this. Even though it doesn't violate any contract, it would be good to cancel the coroutine before executing the listeners.